Use react-icons and component for post header
