+++ package.json
@@ -28,6 +28,7 @@
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-helmet": "^6.1.0",
+    "react-icons": "^4.2.0",
     "tailwindcss-filters": "^3.0.0",
     "typescript": "^4.2.2"
   },
